Switch to generated build TeX file
The new Makefile and associated script generates a container TeX file
for development and building that includes all the diagrams in img-src.
This removes the TeX file from the repo and ensures it will always
include all diagrams.

This approach also moves include and library declaration into the .tikz
files. These declarations are read in during TeX file and PNG file
generation, and the appropriate include directives are emitted into the
TeX files that are eventually compiled.