Configurable figure type
We need a possibility to build diagrams in landscape orientation.
Here we introduce a new metadata property of the tikz file `figuretype`.
For landscape diagrams, it should be set to `sidewaysfigure`, in other
cases to `figure`. We are parsing the value and set it in tex file in
`begin` and `end` tags around the diagram.