Switch to bitcoin-spv special branch
For development purposes we need to switch to a special branch from
bitcoin-spv which contains a feature which will be contained in version
2.0. The error codes returned by headers chain validation will be
changed not to contain values 1,2,3 which seems valid cases for bitcoin
difficulty from a bitcoin testnet perspective. This change is temporary,
we will switch to bitcoin-spv v2.0 when it's released.