Collection of headbanging to get migrations working
This repo presented some new challenges in getting migrations via Circle
working.  The primary offenders are the time it takes to run them and
the composition of package.json.  In package.json there are two private
repos sourced.  To properly source these from CI we would have to create
SSH keys on the utility box inside each VPC where deployments are
applicable.  We need one SSH key for each repo and an SSH / GIT config
to properly select those keys depending on the repo sourced.  These
private repos aren't required for migrations so I'm not doing this work
now.  The other issue is the truffle version.  CI fails to migrate
contracts on the truffle version referenced in package.json.  Pinning
the version to 5.0.7 allows migrations to run.  I've reproduced this
failure on my local machine.