Get a working provisioner InitContainer
We had to do a bit of surgery to retrofit the InitContainer provisioning
scripts to work with tbtc-maintainers.  The most notable bit was in
surfacing the ETH account private key to pass to the tbtc-maintainers
config.  Worth noting that even though this is the tbtc repo, the
InitContainer is for the app that lives in the tbtc-maintainers repo.
We do this because the contract data for tbtc-maintainers lives in tbtc
and we need that data for use in the InitContainer.