Barebones script to provision uniswap contract
Uniswap contracts get migrated from the keep-network/uniswap repo.  When
a migration happens a new contract address is issued for each contract.
The tbtc project needs to have the most recent `UniswapFactoryAddress`
as a reference in its own migrations.  To facilitate this we write
contract migration data from keep-network/uniswap to a bucket designated
for contract data.  Here we pickup that file and parse the needed
contract address.