Set KeepRegistryAddress
externals.js depends on two contract addresses.  We already set the
UniswapFactoryAddress.  Here we add the bits to fetch and set the
KeepRegistryAddress.  We already write compiled contract json for
keep-tecdsa to contract data buckets.  Rather than produce a new
artifact from the keep-tecdsa migration for contract address reading we
use the KeepRegistry.json file already written to storage.  Since we're
leaning on a .json file here, we use jq to parse the needed value.