Function to covert (r,s) signature to bitcoin sig
Extracted code related to conversion of a signature in (r,s) format to
DER format required by bitcoin.
Added check for s value, when s is a high value we convert it to low
value (-s) as per BIP-0062.
Changed import of secp256k1 to a `js` implementation which gives us
access to `curve.n` value.