Updated deps install in Circle CI migrate contracts
Updated the way we install npm dependencies. We used to install them
manually due to a problem with a truffle version we configured in npm.
We updated version of truffle to the one which seems to migrate
contracts correctly in the Circle CI execution. Here we remove a
workaround and replace it with simple `npm install`.