Introduce Circle workflow for keep-test env
Here we provide a gated workflow to migrate contracts and publish data
against the keep-test environment.  We gate via two methods.  The first
filters on git tags to only fire the workflow if a new tag/release is
cut.  The second is a manual approval step to make sure we can control
when the migration happens, as to not disturb any ongoing activity in
the the environment.