Add mock medianizer feeds
The mock medianizer is a simple mock of the official MakerDAO medianizers. It allows a contract owner to update the price of the feed, and adheres to the  Medianizer API for read() only.

Here we deploy two:
- the BTC/USD feed
- the ETH/USD feed

On migration, prices are sourced from a `prices.json` config file.

Two separate contract artifacts simplifies the getting of contract addresses, and mimics the real-life environment.