add RequestorAddress as a manual input on redemption request
Using msg.sedner as the requestor address limits the effectiveness of external redemption calls.
If an external contract redeems on behalf of a user, the redeeming contract must implement extra functionality to
distribute the signer bond in case of fraud during redemption.
It's cleaner to have requestorAddress settable during a redemption request.