Add FundingScript to simplify one-click TBTC minting
FundingScript is a script that mints TBTC from an unqualified DOT. It calls out to the VendingMachine.unqualifiedDepositToTbtc, pre-approving the DOT transfer beforehand using the approveAndCall pattern.

The approach of a separate contract, as opposed to implementing approveAndCall on the vending machine, is best summarised by the Maker team:

"Instead of imposing the design constraints of good end-user ergonomics on the core smart contracts, we move it
to an additional compositional layer of ... stateless scripts." [1]

Ergonomics-wise, this contrasts with the alternative: changing the call semantics of VendingMachine's tbtcToDot and dotToTbtc, to use an additional argument in place of using `msg.sender`.

[1] https://github.com/makerdao/developerguides/blob/master/devtools/working-with-dsproxy/working-with-dsproxy.md