Add deployTestDeposit test helper
For now, only VendingMachineTest moves to it. deployTestDeposit
deploys a common set of contracts, and returns the most important ones
in an object for retrieval by the caller. It allows for deploying more
than the core set of contracts, as well as substituting test or mock
contracts in place of the default ones.

This should let most deposit test files be able to reuse a common
deployment pipeline instead of repeating the same code in each place.