Initialize function for DepositFactoryAuthority
Previously, the authority was initialized in the constructor. Because
deposits are created by cloning a master Deposit, the constructor
parameter can't be correctly passed to the newly-created clone. Here,
the authority is initialized via a separate function, called from the
DepositFactory during new deposit creation.

This failed to be caught because tests used TestDeposit throughout the
test suite. The tests around clone creation now specifically use
Deposit. This is confirmed to fail without the changes in this commit,
and to pass with them.

Finally, the question remains, if the authority is set during
DepositFacotry.createNewDeposit, why have a separate initialize function
instead of removing this ACL (since it only guards
Deposit.createNewDeposit and DepositFunding's own createNewDeposit fails
if it is called a second time)? By doing it this way we still offer
maximum restriction to the state mutation that comes from
createNewDeposit, and limit what an external caller is allowed to do by
API contract, making intended call patterns explicit.