Move npm i and contract compilation to cached build step
There is now a separate build step, compile_contracts, that handles npm
install as well as truffle compile. This step caches the node_modules
across builds, and persists the compiled contracts to the workspace. All
downstream jobs rely on the already-resolved modules and
already-compiled contracts. Currently this includes the utility box
during contract migration, though if there are any native dependencies
that are incompatible between the Circle docker container and the
utility box that might get dicey. TBD.