Refresh tbtc-maintainers provisioning
We've taken a pass across all of our provisioning scripts over the last
few weeks and forgot this one.  Here is a collection of bits that line
provisioning up with new config options in our Kube cluster.  Stuff
worth mentioning here is that we no longer create accounts on init and
instead use an assigned account from Kube.