Back to install npm deps and compiling contracts
We ran a test build on the new setup and migrate_contracts were at the 50
minute mark with SCP'ing data to the utility box before we killed it.
For now we're going back to npm installing and compiling contracts on
the utility box, but we did shift to using npm ci to hopefully help
speed things up.