Remove SPVFraudProof
Because an SPV proof must provide the entire Bitcoin
transaction to the proving smart contract, the Ethereum
block gas limit imposes an upper bound on the size of the transaction in question.
Given that signers can easily spend BTC with a transaction
that is too large to validate by way of SPV proof, this method of fraud proof is unreliable at
best. Deposit owners should instead prove fraud by using an ECDSA fraud proof, which
operates on a hash of the
signed message.