Remove log return values
The return values are not used. On approvedToLog error
revert isntead of returning false.
Returning false offers the abhlity to not halt the
deposit if approvedToLog returns false.
This does not fel like a real perk. if approvedToLog returns
false the event loger is not the deposit contract, and should revert.