Don't explicitly exit
We had this idea to explicity exit on migration completion to prevent an
intermittent build timeout issue.  It turns out that running the
exit results in a failed build, every time.  I suspect it has something
to do with exiting inside of the heredoc, and it not terminating
correctly as a result.  Going to dig into it a bit, but rather than
experiment with different exit scenarios now we're going to remove the
exit to get the builds passing.