Minimum viable Kube config for Bcoin
After some experimenting it looks like we wont be able to get our
already running Bitcoind node working with the relay-maintainer.  This
is due to the maintainer using some RPC method calls that just don't
exist on the Bitcoind API.  Here we provide a Kube config to run a bcoin
node.  Bcoin is a Javascript implementation of Bitcoin and has the api
methods we need.  The Docker image that's published by the maintianers
of Bcoin is two years old.  The image we're using here was build locally
on the latest Bcoin release and uploaded to our container registries.
Not a huge fan but it's what we've got.