Add requestFunderAbort function to deposits
This acts as a signal to the signing group corresponding to a deposit
that failed funding that the depositor would like their UTXO back. This
signal is completely unenforced by the system, as once it is deliverable
the funding process has already failed completely and the signers have
retrieved their bonds. However, it allows for off-chain coordination to
return the funder's UTXO in unusual cases such as a funding transaction
that was unprovable to the contract.