Use receipt block info wherever possible in tests
In many cases tests would perform a block lookup, but rather than couple
tightly to an expectation that the block will or won't change between
transaction mining, the receipt can be used to determine the actual
block that contained a transaction, as well as any metadata about it
like the timestamp the transaction had access to while running.

This commit updates places where the latest block was being looked up,
but instead the receipt's block number could be used.