Set relay contract address via env var
Depending on the network we need either the Relay contract address or
the TestnetRelay contract address.  When we did the original
implementation of this InitContainer we were unaware that TestnetRelay
was a thing.  Rather than get fancy in the InitContainer to
conditionally select which package artifact we use, we're just going to
set the contract address directly in the Kube config via env var.