Allow for price manipulation on Ropsten
When deploying the SatWeiPriceFeed contract on Ropsten, it is now
initialized with the ETHBTCPriceFeedMock, and price feed mock's value
is set to 0 so that it behaves as if inactive. Then, the regular Ropsten
ETHBTC price feed operated by Maker is added, so that it is used while
the mock's value is 0.

When running liquidation test scenarios, the mock's value can be set to
a nonzero value to simulate price changes without changing the Maker
feed on Ropsten.