Add the bones of a deposit state machine tester
The core design is a relatively declarative definition of how to move
from one named state to another, with expectations on each state
transition. Then tests can be defined as paths through the state
machine. The tester exercises each "next state" transition at every step
of the path, allowing greater coverage from a single path test. The idea
is also that failed next state transitions will be able to be specified,
which is to say state transitions that should be impossible.