Separated KeepFactorySelection tests from TBTCSystem tests
One of the advantages of having the code inside of a library is that we
can separate tests and exercise only the library code. This is done
here. Instead of testing the behaviour with the entire TBTCSystem, we
test the library behaviour separately which allows for simpler tests and
better coverage.