Remove check on rebatate distribution
Removing this check should be okay since
distributing the fee rebate happens as the last value
transfer before an end-state.
This also means that calling distributeFeeRebate needs to be done more carefully.

The reason for this change was to allow for the funcding script
to correctly return the fee rebate even when
redeemerAddress == rebateTokenHolder.