Rework state machine runner to chain transitions
Previously, the state machine runner would run all state transitions in
the path and set up tests that would run asynchronously. This made it
difficult/impossible to ensure that the tests associated with each state
would run in the expected context. Snapshots could only be used in some
cases, as all state transitions would run before any tests did.

The base runStatePath function is now purely a non-async setup function
that configures a chain of promises. Each state produces a promise that
will be fulfilled once all of the state's tests run AND THEN its
transition to the next state in the path runs. The full run is then a
chain of those promises, so that at each step the tests run, and then
the transition runs.

An additional benefit is that the entire setup of tests is done
synchronously, which allows it to take place in the context of the
current Mocha test suite. This eliminates the need to pass in the mocha
suite and use private methods to attach tests. Instead, direct calls to
`describe` and `it` work.

Finally, because of all of this setup advantage, the test container is
just a single call to the runner inside a `describe` block's synchronous
definition function. Since the test setup happens synchronously, the end
of the describe block has all tests defined, and the test runner picks
all of them up and runs them in turn, iterating through the promise
chain in the process.