Fix occasional CloneFactoryTest failures
There's an additional layer of cleanup to be had with expectEvent, but
for now capturing the block number *before* clones are created allows
for surefire retrieval of ContractCloneCreated. Before, we were
capturing it *after* awaiting the transaction. This seems to work
sometimes, perhaps even often, though it's not clear exactly why that's
the case---probably something around when the first await returns vs
when ganache mines a new block.