Added import of summa-tx contracts to generate artifacts
We need to import TestnetRelay and OnDemandSPV somewhere to build the
artifacts that are required by the deployment scripts but not
necessarily used on all environments. This was initialy added in
bad43c8d26891f3035d35dc68bd6c2a59c3c1ed9 but we removed it later in
a9dd9fd58ccfe5ab276065131fa95a2d6f80dbfc as the place where it was
imported was quite unfortunate. Here we add the import again to let the
deployment scripts work, but also place it under contract specific for
test usage so we won't mix it in the production contract.