Removed MockMedianizer contract duplicate
The contract was defined in two placed: next to the production contracts
and in test directory. As it's a mock used for local development and
testing environments set up we decided to keep it in the test directory
only.

There were inconsitences between these two contract:
1. `value` type was uint128 or uint256, we decided to keep `uint256` as
this is a type that `IMedianizer` interface defines for the `value`.
2. `ETHBTCPriceFeedMock` was missing in the contract under `test`
directory. We decided to add this contract definition to the
`MockMedianizer.sol` file as it's used in deployment scripts and is
stil required.