Use paid Infura account
Here we switch the relay-maintainer over to our paid Infura account.  I
tried to wire this up to Alchemy but no such luck.  We'll have to sort
that out in a separate PR.

For now I've added an Infura host back to our eth-network-ropsten
Secret, it will only be used by the relay-maintainer.