Add disbursal spec fixtures loader
This file uses asciidoctor-js to read the spec's redemption disbursal
tables and convert them to RedemptionPaymentFixture objects that carry
the information in the table in a test-consumable form.

The spec fixtures loader has types defined in JSDoc and can be
type-checked with TypeScript (though the project continues to have no
TypeScript dependency) in VSCode and similar, and is annotated with a
ts-check tag for this purpose.

A single function is exported, `redemptionPaymentFixturesFor`, which
takes a lot size and signer fee and computes the disbursal fixtures with
those values.