Rework redemption payment calculations one more time
Superficially, getRedemptionTbtcRequirement and
getOwnerRedemptionTbtcRequirement once again return only the payment in
TBTC needed from the redeemer. Redemption no longer pays anything out to
the redeemer, ever. Any excess amounts in the deposit's balance when
redemption is initiated is returned to the TDT holder.

The second major factoring change is renaming
getRedemptionTbtcRequirement to calculateRedemptionTbtcAmounts, to
reflect the fact that it returns both amounts owed to the deposit and
outwards to other parties. The logic of calculateRedemptionTbtcAmounts
is also completely refactored, and is fully reused from
getRedemptionTbtcRequirement and getOwnerRedemptionTbtcRequirement to
minimize opportunity for divergence. This is achieved by adding a
parameter, _assumeRedeemerHoldsTdt, that acts as the sole override
needed by getOwnerRedemptionTbtcRequirement.

The fee escrow requirement, including FRT rebate, is computed in one
place now (computeRedemptionFeeEscrow). The body of
calculateRedemptionTbtcAmounts is simplified to first compute the total
escrow needed for the deposit, then compute how much the redeemer has to
pay to reach that escrow requirement, and finally compute the amounts
owed to the FRT and TDT holders based on the total deposit balance.
Any balance in the deposit counts towards the redemption TBTC
requirement, and any leftover balance after redemption payouts and the
escrow for signers (who are paid at the end of the redemption flow
rather than the beginning).