Limit initial redemption fee to <1/2 of deposit value
In combination with an adjustment to the max fee bump, this will prevent
fees that will fail the dust rule on the Bitcoin side from being
requested.

In particular, Bitcoin relay rules generally reject transactions from
the mempool if the value of a UTXO they create is less than a certain
amount based on various rules. An initial fee request of this size would
allow a malicious redeemer to force the signers to sign a transaction
that can't be mined without asking a miner to directly include the
transaction. This would most likely lead to a scenario where the signers
time out the redemption proof and their bonds are seized and sent to the
malicious redeemer. In addition to the net-positive-value outcome if the
redeemer has a direct relationship with a miner (or is themselves a
miner), they can then mine the transaction and also gain the underlying
BTC.

Limiting the initial requested fee to less than half of the deposit
value ensures the first transaction will not produce a dust output.
There is still the potential for this fee to be too _low_ if there is an
ongoing fee event, however---that will be addressed by a separate
change.

A few adjustments needed to be made to the fixtures used throughout
redemption and vending machine tests, which largely used too high a fee.
Expected sighashes were also revised accordingly.