Prevent redemption UTXO from dipping below a minimum value
Tags along with the previous commit, which limits the starting value of
the redemption fee, to  prevent fees that will fail the dust rule on the
Bitcoin side from being requested.

In particular, Bitcoin relay rules generally reject transactions from
the mempool if the value of a UTXO they create is less than a certain
amount based on various rules. Fee bumps can currently bump the fee to a
high enough value that a malicious redeemer can force the signers to
sign a transaction that can't be mined without asking a miner to
directly include the transaction. This would most likely lead to a
scenario where the signers time out the redemption proof and their bonds
are seized and sent to the malicious redeemer. In addition to the
net-positive-value outcome if the redeemer has a direct relationship
with a miner (or is themselves a miner), they can then mine the
transaction and also gain the underlying BTC.

Setting a minimum UTXO value (set in TBTCConstants) ensures a fee bump
will not produce a dust output. The bumping process is also adjusted so
that, if the final fee bump pushes the UTXO value below the minimum,
the fee bump is permitted but clamped to a value that leaves the minimum
UTXO value intact.