Create sortition pool via Circle
We currently create sortition pools for new TBTCSystem applications
during the keep-ecda application deployment init process.  While this
works the creation is misplaced.  Once a sortition pool is created for
an application that's it for the application, we won't create another
one.  This translates to the very first keepe-ecdsa  client  deployed
against a new TBTCSystem creating a pool, and all subsequent deployments
wasting resources on the sorition pool check.

Instead were going to run sortition pool creation as part of the tbtc
Circle migration flow.  After tbtc migrations are finished we'll run
this script in a Circle job to create the sortition pool.  We'll remove
the sortition pool check and create logic from the keep-ecdsa
InitContainers.