Reorganize functions in Deposit contract
The flow of the contract file now starts with state reading functions,
followed by mutating functions in order of the flows that deposits go
through; namely funding -> liquidation -> redemption. At the end are a
handful of mutating helpers that don't clearly belong to any of the
specific flows or are public but have caller restrictions that make them
external to the main deposit flows.