Refresh the minimum bondable value for an operator
Deep inside ECDSA keep factory implementation, `BondedSortitionPool` was
updating the minimum bondable value for each new group selection. The idea
was to automatically remove from the pool operators that are no longer
able to satisfy application needs about the minimum available bondable value
and prevent operators with too low value griefing signer selection.

We now have two minimum bond values: one defining the minimum unbonded value
the operator needs to have so that it can join and stay in the pool, and
another defining the required bond value for the current selection. If the
given operator has enough unbonded value to stay in the pool but it does not
have enough unbonded value for the current group selection, it is skipped.

In ECDSA keep factory, sortition pool is created with a minimum bond of 20 ETH
to avoid small operators joining and griefing future selections before the
minimum bond is set to the right value by the application. Anyone can create
a sortition pool for an application with the default minimum bond value but
the application can change this value later, at any point. 20 ETH per operator
should be more than enough to cover 1 BTC deposit in 3-operators group.

The default is fine for most cases <now> but the application is expected to
refresh the minimum bondable value, depending on the current situation and
needs.

For tBTC, the minimum bond value proposed here is be the minimum lot size
expressed in wei given the current BTC price with 150% collateralization.
For such a value, operator should be able to participate in at least 3 keeps,
or - in extreme circumstances - hold one keep just by themselves.

Here we add a function allowing to refresh the minimum bondable value given
the current BTC price. The function can be called by anyone at any moment.
It is recommended to call this function on tBTC initialization and after
the minimum lot size update.