Create sortition pool
We currently create sortition pools for new TBTCSystem applications
during the keep-ecda application deployment init process.  While this
works the creation is misplaced.  Once a sortition pool is created for
an application that's it for the application, we won't create another
one.  This translates to the very first keepe-ecdsa  client  deployed
against a new TBTCSystem creating a pool, and all subsequent deployments
wasting resources on the sorition pool check.

Instead were going to run sortition pool creation as part of the tbtc
migration flow.  We'll remove the sortition pool check and create logic
from the keep-ecdsa InitContainers.