Keep group size and threshold values in TBTCSystem
This way, we do not need to pass those values around from DepositFactory
to DepositFunding and Deposit, and have them eventually passed to
TBTCSystem. Group size and threshold are global, system parameters so it
does make sense to keep them on a global level. Also, this change will
help us clean up bond calculation code (more to come in the next
commits).