Properly burn TDT holder owed amount at redemption
The amount being burned was the amount owed to the deposit instead of to
the TDT holder when the Vending Machine held the TDT. This was related
to the fee refactor, which added new return values to
calculateRedemptionTbtcAmounts, but did not update the location of the
TBTC burn. Burning now happens in the block that handles
tbtcOwedToTdtHolder, and burns only from the deposit's balance.