Switch public to external where possible
There are a few cases where a public to external switch *would* be
possible, if it weren't for the fact that the code requires parameters
that must be passed in as `bytes memory`. In these cases, the function
is marked with a comment indicating the reason for the choice.

In certain cases, some light reordering of functions was required to
conform to the external-then-public standard function ordering. In some
cases, this lint rule was disabled as the public functions were so
marked only due to requiring `bytes memory` parameters, and removing
them from their place would affect overall file organization.