Obtain Keep factories addresses through vendor
Keep ECDSA defines a Vendor contract that is used to register factories
and returns the latest upgraded version of a factory. We were using the
vendor contract previously but decided to use a fixed factory address.
With this change we are adding back usage of vendor contract. The
contract will be called to obtain the latest version of a Keep factory
for each type: KEEP staked and ETH only backed keeps.