Added lock for Keep factories version updates
Here we are adding locking mechanism for Keep factories version updates.
We have a possibility to update factory address via keep vendor
contract. We want to have a way to lock the versions at some point in
the future when keep-ecdsa is considered to be in a state that won't
require any updates. When lock is set we will use only the versions of
the factories that were current ones in the moment of locking.