Don't store factory address on setFullyBackedKeepVendor
We don't store factory address that is returned from a vendor on
setFullyBackedKeepVendor call. We don't need this value stored as on
refresh we're calling the vendor to get the latest one. This saves some
gas. Though, we still call the vendor to return a factory to verify that
it's able to return an address on the `selectFactory` call and that the
address is set to non-default value other than `0`.