Updated setMinimumBondableValue function for factories
Since we added vendors and are not storing the factories until they are
locked we had to update the setMinimumBondableValue function to use the
same function as refresh to obtain factories addresses.
It is expected that KEEP staked vendor is set at the very begining with
default initialization.
Fully backed vendor is not required for this function to complete.