Allow multiple updates of fully backed keep factory
We allowed multiple updates of a fully backed keep factory and factory
selector contracts addresses. The limitation is that updates can be
initiated only within a upgreadability period. Once upgradeability
period passes factories versions are freezed and no change is allowed.

We updated naming to remove "SingleShot", as updates are now allowed to
happen multiple times until they are freezed.