Allow zero addresses for fully backed factory and factory selector
We removed a requirement that fully backed factory and factory selector
addresses on update have to be different than zero.

This lets us to initiate an update where when fully backed bonding is
not ready yet we're updating just a KEEP-staked factory.

The other possibility is that once fully backed bonding is enabled an
update can be initiated that disables this functionality to just
KEEP-staked keeps creation.

KEEP-staked factory address is still required to be other than zero as
it always have to be valid for the system to be working as a default
choice for the factory.