Do npm ci before npm update
This is a temporary fix to avoid problems with pulling non-existing
dependency on some bignumber.js fork:

npm ERR! Could not install from "node_modules/web3/bignumber.js@git+
https:/github.com/debris/bignumber.js.git#94d7146671b9719e00a09c29b0
1a691bc85048c2" as it does not contain a package.json file.

Ironically, it all works fine if we do `npm ci` first.

The proper solution would be (probably) to remove no longer used
`solidity-bytes-utils` from `keep-ecdsa` and updating the dependency
here.

Given that we are in the mid of deploy and this dependency is not used,
we are going to do a workaround for now and implement a proper solution
later.