Resolve latest dependencies for keep-test at begining
We don't want to update dependencies version right before package
publication as all the steps were executed with specific locks on
dependencies.
Updating dependencies version should be made at earlier stage, tested
and reflected in the package.json and package-lock.json we've got in the
repository.

For keep-test migration we need to update versions of packages we depend
on `@keep-network/keep-ecdsa`. With this change we extracted this part
to a separate step that is executed at the very begining of the process.
With a properly configured `package.json` semver for `keep-ecdsa` we
will resolve latest version before compilation and migration of
contracts.