Refresh minimum bondable value after sortition pool creation
Sortition pool gets created with a default minimum bondable value
defined in keep factory (at the moment it's 20 ETH).
TBTCSystem contract is responsible of updating the value to the actually
required.
Here we call `refreshMinimumBondableValue` function to configure the
value for the newly created pool.

This is required for our CI to lower requirements for operators joining the pool.