Remove node_modules caching
We removed steps related to node_modules caching as this was causing
problems when `compile_contracts` was used concurrently in multiple
jobs.
Also there is no point in caching `node_modules` when we use `npm ci` as
the command does clean install by removing `node_modules` at the very
begining of the execution.