Resolve latest dependencies for keep-test at begining
For keep-test migration we need to update versions of packages we depend
on `@keep-network/keep-ecdsa`. With this change we extracted this part
to a separate step that is executed at the very begining of the process.
With a properly configured `package.json` semver for `keep-ecdsa` we
will resolve latest version before compilation and migration of
contracts.