Create GH Actions workflow for lint of Solidity contracts
As described in RFC-18, there is a need for a refactorization of Keep
and tBTC release processes in order to reduce human involvement and make
the processes faster and less error prone. A part of this task is
migration from CircleCI jobs to GitHub Actions. This commit creates a
GitHub Action workflow for linting of Solidity contracts.