Create GH Actions workflow for build&test of Solidity contracts
As described in RFC-18, there is a need for a refactorization of Keep
and tBTC release processes in order to reduce human involvement and make
the process less error prone. A part of this task is migration from
CircleCI to GitHub Actions. This commit creates a GitHub Action workflow
for building and testing of Solidity contracts.