Move migrate&publish of Eth contracts to GH Actions
As described in RFC-18, there is a need for a refactorization of Keep
and tBTC release processes in order to reduce human involvement and to
make the processes faster and less error prone. A part of this task is
migration from CircleCI jobs to GitHub Actions. This commit enhances
the workflow for building and testing Solidity with a job performing
contracts migration on Ethereum chain.
The job does following:
- fetches the contracts artifacts uploaded to Google Storage bucket by
  Solidity workflow in`keep-ecdsa`
- migrates contracts
- pushes them to Tenderly
- uploads contract data to Google Storage bucket (`tbtc`)
- bumps up version of npm package containing contract artifacts and
  uploads the package to npm registry

The `migrate-and-publish-ethereum` job is configured to execute only if
workflow is started by a push or workflow_dispatch event on master or
`rfc-18/*` branch. The `rfc-18/*` condition will need to be removed
once whole `build-test-migrate-publish-keep-test` CircleCI process gets
migrated to GH Actions.
Additionally, as we may not always want to migrate and publish
contracts on every push (or workflow_dispatch) on master, the job was
configured to require manual approval in GitHub GUI. This is done by
configuration of environment protection rules on keep-test and using
`environment: keep-test` config in the workflow.