Cleanup after moving most jobs to GH Actions
As described in RFC-18, there is a need for a refactorization of Keep
and tBTC release processes in order to reduce human involvement and
make the processes faster and less error prone. A part of this task is
migration from CircleCI jobs to GitHub Actions.

Now that we have most CircleCI jobs from `tbtc` moved to GH Actions
(and verified), we can implement some changes removing obsolete
CircleCI configs and cleaning up pieces of code used for workflows
testing. Here is a list of changes applied:
- removed unnecessary comments
- removed references to `rfc-18...` branches, which were used for
  workflow testing
- removed `--dry-run` flag from command publishing npm image
- deleted couple of CircleCI jobs from `./circleci/config.yml` (those
  who are migrated to GH Actions and are not necessary for building of
  relay maintainer initcontainer)