Test notify-workflow-completed action
As part of work on RFC-18 a mechanism for managing triggering of
workflows has been created. A part of this mechanism is
the `notify-workflow-action` which will be executed at the end of the
workflow and will notify the release manager repository about build
completion.
This commit introduces the notify-workflow-completed action to
`contracts.yml` workflow in order to test if the communication between
particular components of the mechanism works properly.