Update version of tbtc package
To avoid mixup of versions of migrated contracts test and dev
environments must operate on different versions of `tbtc` package.
Packages with `-rc` suffix will be used on test env (Ropsten). Packages
with `-pre` suffix will be used for development purposes.
The commit modifies the version of `tbtc` package and updates the
versions of `keep-network/...` dependencies.