Replace the CircleCI build status badge with GHA badges
We no longer use CircleCI to test/build the `tbtc` code. We need to
replace the CircleCI build status badge in the projec'ts README with
the similiar badge for GitHub Actions build.
The badge has been replaced with the Solidity build badge (showing
status of the last `Solidity` workflow executed on `master` branch and
triggered by the push event).
The badges is displayed using `shields.io` tool, which was already
used for the Discord and docs badges.