Move job building&publishing Relay Maintainer Initcontainer to GH Actions
As described in RFC-18, there is a need for a refactorization of Keep
and tBTC release processes in order to reduce human involvement and
make the processes faster and less error prone. A part of this task is
migration from CircleCI jobs to GitHub Actions. This commit enhances
the Solidity workflow with a job building Relay Maintainer
Initcontainer and publishing it to the Google Container Registry.

Temporarily, the name of the published image ends with `-wip` suffix.
This was added only for testing purposes and will be deleted before
submitting final PR.
Also, the `envronment: keep-test` line was temporarily commented out for
testing purposes. Although, generally the line is no longer needed, it
will be brought back in one of the next commits (it's going to be
deleted in other PR).