Move initcontainer to the Relay workflow
As the job building and publishing relay maintainer initcontainer will
is not using artifacts of other jobs/workflows, it does not need to be
executed as part of flow deploying on testnet. This means it does not
need to be a part of `contracts.yml` workflow (which is executed in
that flow). We're moving the job to `relay.yml` workflow. Job will be
executed if workflow will be triggered manually (via
`workflow_dispatch` event).