Add keep-prd compact history job for ElectumX
ElectrumX instances can enter a state where their index is too large and
requires out-of-band compacting, at which point the ElectrumX server
will crash and fail to retart. The job added here can be run once and
will compact the history.

It needs to be pointed to the correct volume via the
persistentVolumeClaim, and is designed to run only once to
completion---successful or not. To run again later, the job can be
deleted from the GCP UI or using kubectl, and then re-applied.